requirements.txt
ansi2html==1.9.1
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-MySQLdb==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
pexpect==4.9.0
psutil==5.9.8
ptyprocess==0.7.0
pytz==2024.1
Werkzeug==3.0.1
